# docs/api/hasivu-platform-openapi.yaml
openapi: 3.0.3
info:
  title: HASIVU Platform API
  description: |
    Complete API documentation for the HASIVU Platform - A comprehensive school meal delivery system with AWS Lambda-based microservices architecture.

    ## Features
    - JWT-based authentication with AWS Cognito integration
    - Payment processing with Razorpay integration
    - RFID-based delivery verification
    - Real-time notifications via WhatsApp
    - Advanced analytics and reporting
    - Subscription billing management
    - AI-powered payment intelligence

    ## Authentication
    This API uses JWT tokens for authentication. Include the JWT token in the Authorization header:
    ```
    Authorization: Bearer <jwt_token>
    ```

    ## Base URL
    - **Production**: `https://api.hasivu.com`
    - **Staging**: `https://api-staging.hasivu.com`
    - **Development**: `https://api-dev.hasivu.com`

    ## Rate Limiting
    - **Rate Limit**: 1000 requests per minute per IP
    - **Burst Limit**: 2000 requests per minute

    ## Error Handling
    All errors follow RFC 7807 problem details format with consistent structure.
  version: 1.0.0
  contact:
    name: HASIVU API Support
    email: api-support@hasivu.com
    url: https://hasivu.com/support
  license:
    name: Proprietary
    url: https://hasivu.com/terms
  termsOfService: https://hasivu.com/terms
servers:
  - url: https://api.hasivu.com
    description: Production server
  - url: https://api-staging.hasivu.com
...
security:
  - bearerAuth: []
paths:
...
  /rfid/readers/{readerId}/test:
    post:
      tags:
        - RFID
      summary: Test RFID reader
      description: Test RFID reader connection
      operationId: testRfidReader
      security:
        - bearerAuth: []
      parameters:
        - name: readerId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: RFID reader test completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RfidTestResponse'
...
  /invoices/generate:
    post:
      tags:
        - Invoices
      summary: Generate invoice
      description: Generate invoice for payments
      operationId: generateInvoice
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenerateInvoiceRequest'
      responses:
        201:
          description: Invoice generated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvoiceResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login endpoint
  schemas:
...
    RfidTestResponse:
      type: object
      properties:
        readerId:
          type: string
          format: uuid
        status:
          type: string
          enum: [success, failure, timeout]
        responseTime:
          type: integer
          description: Response time in milliseconds
        errorMessage:
          type: string
          nullable: true
        testedAt:
          type: string
          format: date-time
...
    GenerateInvoiceRequest:
      type: object
      required:
        - userId
        - items
      properties:
        userId:
          type: string
          format: uuid
        items:
          type: array
          items:
            type: object
            properties:
              description:
                type: string
              quantity:
                type: integer
                minimum: 1
              unitPrice:
                type: number
                format: double
              itemType:
                type: string
                enum: [meal, delivery_fee, service_charge]
        dueDate:
          type: string
          format: date
        templateId:
          type: string
          format: uuid
          nullable: true
    InvoiceResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        invoiceNumber:
          type: string
        invoiceDate:
          type: string
          format: date
        dueDate:
          type: string
          format: date
        subtotal:
          type: number
          format: double
        taxAmount:
          type: number
          format: double
        totalAmount:
          type: number
          format: double
        currency:
          type: string
        status:
          type: string
          enum: [draft, sent, paid, overdue, cancelled]
        pdfUrl:
          type: string
          nullable: true
        emailSent:
          type: boolean
        items:
          type: array
          items:
            type: object
            properties:
              description:
                type: string
              quantity:
                type: integer
              unitPrice:
                type: number
                format: double
              totalPrice:
                type: number
                format: double
              itemType:
                type: string
...